3. Entity 변환

+++ src/test/java/com/makers/princemaker/service/PrinceMakerServiceTest.java
@@ -77,7 +77,7 @@ class PrinceMakerServiceTest {
 		Prince savedPrince = captor.getValue();
 		assertEquals(MIDDLE_PRINCE, savedPrince.getPrinceLevel());
 		assertEquals(INTELLECTUAL, savedPrince.getSkillType());
-		assertEquals(7, savedPrince.getExperienceYears().intValue());
+		assertEquals(7, savedPrince.getExperienceYears());
 
 		assertEquals(MIDDLE_PRINCE, response.getPrinceLevel());
 		assertEquals(INTELLECTUAL, response.getSkillType());
